backend/requirements .txt
#LangChain core 
LangChain
langchain-core

#Google Gemini (PaLM) Integaration
langchain-google-genai 
google-generativeai

#Hugging Face Integaration
langchain-huggingface
transformers
huggingface-hub

#Environment Variables MAnagement
python-dotenv

# LangChain Core
langchain>=0.1.0
langchain-google-genai>=0.0.8  # Official LangChain wrapper for Gemini
python-dotenv>=1.0.0

# Google Generative AI SDK
google-generativeai>=0.3.2

# Resume file extraction
PyMuPDF==1.23.2         # for PDF
python-docx==1.1.0      # for DOCX

# Optional grammar tools
language-tool-python==2.7.1




